# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "Blog post",
    "description": "Blog post management API",
    "termsOfService": "",
    "contact": {
      "name": "M2M Team",
      "email": "team@mind2minds.com",
      "url": ""
    },
    "license": {
      "name": "Apache 2.0",
      "url": ""
    },
    "securityDefinitions": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "paths": {
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "description": "Get users",
        "operationId": "getUsers",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/token"
            },
            "required": "true",
            "description": "User login token"
          }
        ],
        "responses": {
          "200": {
            "description": "List of all users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              }
            }
          },
          "400": {
            "description": "Missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "status": "error",
                  "message": "Invalid token"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "description": "Create users",
        "operationId": "createUsers",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/token"
            },
            "required": "true",
            "description": "User login token"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUsers"
              }
            }
          },
          "required": "true"
        },
        "responses": {
          "200": {
            "description": "New user were created successfully"
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "status": "error",
                  "message": "User already exist"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": ["Users"],
        "description": "Get user by id",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/token"
            },
            "required": "true",
            "description": "User login token"
          },
          {
            "name": "id",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Founded user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              }
            }
          },
          "400": {
            "description": "Missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "status": "error",
                  "message": "Invalid token"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "description": "Update user",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/token"
            },
            "required": "true",
            "description": "User login token"
          },
          {
            "name": "id",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUsers"
              }
            }
          },
          "required": "true"
        },
        "responses": {
          "200": {
            "description": "Updated users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              }
            }
          },
          "400": {
            "description": "Missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "status": "error",
                  "message": "Invalid user"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "description": "Delete user",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/token"
            },
            "required": "true",
            "description": "User login token"
          },
          {
            "name": "id",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User is deleted successfully"
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "status": "error",
                  "message": "Invalid token"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/search": {
      "get": {
        "tags": ["Users"],
        "description": "Search users",
        "operationId": "searchUser",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/token"
            },
            "required": "true",
            "description": "User login token"
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              }
            }
          },
          "400": {
            "description": "Missing parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "status": "error",
                  "message": "Invalid token"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/sign-in": {
      "post": {
        "tags": ["Users"],
        "description": "Sign-in user",
        "operationId": "SignInUser",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInUsers"
              }
            }
          },
          "required": "true"
        },
        "responses": {
          "200": {
            "description": "User Sing-in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/token"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "status": "error",
                  "message": "Invalid user credentials"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Authorization": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "id": {
        "type": "integer",
        "description": "unique identification number",
        "example": "1234"
      },
      "firstName": {
        "type": "string",
        "description": "First name of the user",
        "example": "Jon"
      },
      "lastName": {
        "type": "string",
        "description": "Last name of the user",
        "example": "doe"
      },
      "email": {
        "type": "string",
        "description": "email of the user",
        "example": "test123@test.com"
      },
      "password": {
        "type": "string"
      },
      "profileImageUrl": {
        "type": "string"
      },
      "lastLoggedIn": {
        "type": "string"
      },
      "roleId": {
        "type": "integer"
      },
...
      "Users": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/id"
          },
          "firstName": {
            "$ref": "#/components/schemas/firstName"
          },
          "lastName": {
            "$ref": "#/components/schemas/lastName"
          },
          "email": {
            "$ref": "#/components/schemas/email"
          },
          "password": {
            "$ref": "#/components/schemas/password"
          },
          "profileImageUrl": {
            "$ref": "#/components/schemas/profileImageUrl"
          },
          "lastLoggedIn": {
            "$ref": "#/components/schemas/lastLoggedIn"
          },
          "roleId": {
            "$ref": "#/components/schemas/roleId"
          }
        }
      },
      "CreateUsers": {
        "type": "object",
        "properties": {
          "firstName": {
            "$ref": "#/components/schemas/firstName"
          },
          "lastName": {
            "$ref": "#/components/schemas/lastName"
          },
          "email": {
            "$ref": "#/components/schemas/email"
          },
          "password": {
            "$ref": "#/components/schemas/password"
          },
          "profileImageUrl": {
            "$ref": "#/components/schemas/profileImageUrl"
          },
          "roleId": {
            "$ref": "#/components/schemas/roleId"
          }
        }
      },
      "UpdateUsers": {
        "type": "object",
        "properties": {
          "firstName": {
            "$ref": "#/components/schemas/firstName"
          },
          "lastName": {
            "$ref": "#/components/schemas/lastName"
          },
          "email": {
            "$ref": "#/components/schemas/email"
          },
          "password": {
            "$ref": "#/components/schemas/password"
          },
          "profileImageUrl": {
            "$ref": "#/components/schemas/profileImageUrl"
          },
          "lastLoggedIn": {
            "$ref": "#/components/schemas/lastLoggedIn"
          },
          "roleId": {
            "$ref": "#/components/schemas/roleId"
          }
        }
      },
      "SignInUsers": {
        "type": "object",
        "properties": {
          "email": {
            "$ref": "#/components/schemas/email"
          },
          "password": {
            "$ref": "#/components/schemas/password"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "msg": {
            "type": "string"
          }
        }
      }
    }
  }
}